Updated server structure.
