--- .env.sample
@@ -1,6 +1,0 @@
-# View README.md for more information
-export AUTH_CLIENT_ID=
-export AUTH_CLIENT_SECRET=
-export AUTH_SUBSCRIPTION_KEY=
-export AUTH_REDIRECT_URI=https://localhost:5000/auth/callback
-export PORT=5000
