+++ .gitignore
@@ -2,8 +2,5 @@
 npm-debug.log
 node_modules
 bower_components
-.env
-!.env.sample
-env.bat
-!envsample.bat
+*.env
 .DS_Store
