+++ index.js
@@ -1,91 +1,69 @@
 /*jslint node: true, es5: true, nomen: true*/
-'use strict';
+(function () {
+    'use strict';
 
-var auth,
-    apiNxt,
-    apiSky,
-    app,
-    bodyParser,
-    express,
-    fs,
-    http,
-    https,
-    httpsOptions,
-    session,
-    sessionConfig,
-    timeout;
+    var app,
+        bodyParser,
+        express,
+        fs,
+        http,
+        https,
+        httpsOptions,
+        routes,
+        session,
+        sessionConfig,
+        timeout;
 
+    // Application dependencies
+    fs = require('fs');
+    http = require('http');
+    https = require('https');
+    routes = require('./server/routes');
+    express = require('express');
+    session = require('express-session');
+    timeout = require('connect-timeout');
+    bodyParser = require('body-parser');
 
-// Application dependencies
+    // If deployed in our demo site, we store the sessions using Redis
+    // Locally we store the sessions in memory
+    sessionConfig = {
+        resave: false,
+        saveUninitialized: true,
+        secret: '+rEchas&-wub24dR'
+    };
 
-auth = require('./server/auth.js')();
-apiNxt = require('./server/api-nxt.js')(auth);
-apiSky = require('./server/api-sky.js')(apiNxt);
-bodyParser = require('body-parser');
-express = require('express');
-fs = require('fs');
-http = require('http');
-https = require('https');
-session = require('express-session');
-timeout = require('connect-timeout');
+    // Create our application and register its dependencies
+    app = express();
+    app.use(bodyParser.json());
+    app.use(session(sessionConfig));
+    app.use(timeout('30s'));
 
-// If deployed in our demo site, we store the sessions using Redis
-// Locally we store the sessions in memory
-sessionConfig = {
-    resave: false,
-    saveUninitialized: true,
-    secret: '+rEchas&-wub24dR'
-};
+    // Register our OAUTH2 routes
+    app.get('/auth/authenticated', routes.auth.getAuthenticated);
+    app.get('/auth/login', routes.auth.getLogin);
+    app.get('/auth/callback', routes.auth.getCallback);
+    app.get('/auth/logout', routes.auth.getLogout);
 
-// Create our application and register its dependencies
-app = express();
-app.use(bodyParser.json());
-app.use(session(sessionConfig));
-app.use(timeout('30s'));
+    // Register our SKY API routes
+    app.get('/api/constituents/:constituentId', routes.auth.checkSession, routes.api.sky.getConstituent);
 
-// Register our OAUTH2 routes
-app.get('/auth/authenticated', auth.getAuthenticated);
-app.get('/auth/login', auth.getLogin);
-app.get('/auth/callback', auth.getCallback);
-app.get('/auth/logout', auth.getLogout);
+    // Register our front-end UI routes
+    app.use('/', express.static(__dirname + '/ui'));
 
-// Validates all requests
-function requireSession(request, response, next) {
-    auth.validate(request, function (valid) {
-        if (valid) {
-            next();
-        } else {
-            response.sendStatus(401);
-        }
+    app.get('/', routes.auth.checkSession, function (request, response) {
+        response.json({
+            access_token: request.session.ticket
+        });
     });
-}
 
+    // Displays the startup message
+    function onListen() {
+        console.log('SKY API Auth Code Flow Tutorial app running for https://localhost:%s/', process.env.PORT);
+    }
 
-
-// Register our SKY API routes
-app.get('/api/constituents/:constituentId', requireSession, apiSky.getConstituent);
-
-
-// Register our front-end UI routes
-//app.use('/', express.static(__dirname + '/ui'));
-app.use('/', express['static'](__dirname + '/ui'));
-
-app.get('/', requireSession, function (request, response) {
-    console.log(request.session.ticket);
-    response.json({
-        access_token: request.session.ticket
-    });
-});
-
-
-
-// Displays the startup message
-function onListen() {
-    console.log('SKY API Auth Code Flow Tutorial app running for https://localhost:%s/', process.env.PORT);
-}
-
-httpsOptions = {
-    key: fs.readFileSync('sslcerts/server.key', 'utf8'),
-    cert: fs.readFileSync('sslcerts/server.crt', 'utf8')
-};
-https.createServer(httpsOptions, app).listen(process.env.PORT, onListen);
+    httpsOptions = {
+        key: fs.readFileSync('./server/sslcerts/server.key', 'utf8'),
+        cert: fs.readFileSync('./server/sslcerts/server.crt', 'utf8')
+    };
+    https.createServer(httpsOptions, app).listen(process.env.PORT, onListen);
+}());
