+++ package.json
@@ -5,13 +5,15 @@
   "main": "index.js",
   "scripts": {
     "build": "grunt build",
-    "start": "node index.js"
+    "start": "node ./bin/start.js"
   },
   "author": "Blackbaud-TripOtt",
   "license": "MIT",
   "dependencies": {
     "body-parser": "^1.14.1",
     "connect-timeout": "^1.7.0",
+    "continuation-local-storage": "^3.1.6",
+    "dotenv": "^2.0.0",
     "express": "^4.13.3",
     "express-session": "^1.11.3",
     "request": "^2.64.0",
