+++ server/sslcerts/server.crt
@@ -1,0 +1,22 @@
+-----BEGIN CERTIFICATE-----
+MIIDtDCCApwCCQDMXIgzspIcjTANBgkqhkiG9w0BAQUFADCBmzELMAkGA1UEBhMC
+VVMxFzAVBgNVBAgTDlNvdXRoLUNhcm9saW5hMRMwEQYDVQQHEwpDaGFybGVzdG9u
+MRIwEAYDVQQKEwlCbGFja2JhdWQxDDAKBgNVBAsTA1JETzESMBAGA1UEAxMJbG9j
+YWxob3N0MSgwJgYJKoZIhvcNAQkBFhlzdGV2ZS5icnVzaEBibGFja2JhdWQuY29t
+MB4XDTE2MDIxODE2MjUwOFoXDTE3MDIxNzE2MjUwOFowgZsxCzAJBgNVBAYTAlVT
+MRcwFQYDVQQIEw5Tb3V0aC1DYXJvbGluYTETMBEGA1UEBxMKQ2hhcmxlc3RvbjES
+MBAGA1UEChMJQmxhY2tiYXVkMQwwCgYDVQQLEwNSRE8xEjAQBgNVBAMTCWxvY2Fs
+aG9zdDEoMCYGCSqGSIb3DQEJARYZc3RldmUuYnJ1c2hAYmxhY2tiYXVkLmNvbTCC
+ASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAKg52W85M/Jt9RnQO/yYce0k
+xqtKK8KvaTIlD/ZlbwjaWVYrr6mWXrAgQz98exCtAvB0qtN6CCvFI91jAgqmq7Wy
+tzsEHY20/OouoZTRUF0odabCFnvJp3Cr77Xgs0+NL1nvorlIxsmWfH6PtxcWW/hq
+e7TLT/mrn8+DJbDrTl4Kmtn0wMdygH+naKO2vU8sgQn8pjSvN8mL6kir5voLc0q7
+/hi69+BfE7Q1smTkGXLsTlSyDSdt+Cc7OdL3FcsctzKemnhaU/cF36QxJdtLg9Av
+OMXW9dyyVQgjUyBSBm5Q7WQvxvo20xlTqCznX7H3ekeRJFIkGN0ns32e4sPjFdkC
+AwEAATANBgkqhkiG9w0BAQUFAAOCAQEAYXV+6+eMZADNf6G6EPdO3QxFTkx/gcVk
+yphWQFjRM+5fUiEWnVzuWBkJlmXaWokaTSgjQKZTnh7vyKVnZQhIO9LBjo4O0Nmo
+3tm0nsrid1bf7LjZWN7XaDi9uCeL7r0mqLkRvhVyPu3JkQya2+MVkT7qTbEjt6la
+32SrAfAJDNbVltzSS2qKR6GQ1JFpjwF22I+OkWe2ZvMqoiHVG1DMPBEiZe0TCHJC
+9J5Rj1wuJ9t1jNJhIGgw7g6yZoPMukJW+mm8kbIzWKTe+KrCXSDIjs198fUQWK/U
+KOT+Q8fiiFN0sN9OFXJz/qQ1dcm9l2427MsPku90/kJz5CAChkCsEQ==
+-----END CERTIFICATE-----
