+++ server/sslcerts/server.csr
@@ -1,0 +1,19 @@
+-----BEGIN CERTIFICATE REQUEST-----
+MIIDETCCAfkCAQAwgZsxCzAJBgNVBAYTAlVTMRcwFQYDVQQIEw5Tb3V0aC1DYXJv
+bGluYTETMBEGA1UEBxMKQ2hhcmxlc3RvbjESMBAGA1UEChMJQmxhY2tiYXVkMQww
+CgYDVQQLEwNSRE8xEjAQBgNVBAMTCWxvY2FsaG9zdDEoMCYGCSqGSIb3DQEJARYZ
+c3RldmUuYnJ1c2hAYmxhY2tiYXVkLmNvbTCCASIwDQYJKoZIhvcNAQEBBQADggEP
+ADCCAQoCggEBAKg52W85M/Jt9RnQO/yYce0kxqtKK8KvaTIlD/ZlbwjaWVYrr6mW
+XrAgQz98exCtAvB0qtN6CCvFI91jAgqmq7WytzsEHY20/OouoZTRUF0odabCFnvJ
+p3Cr77Xgs0+NL1nvorlIxsmWfH6PtxcWW/hqe7TLT/mrn8+DJbDrTl4Kmtn0wMdy
+gH+naKO2vU8sgQn8pjSvN8mL6kir5voLc0q7/hi69+BfE7Q1smTkGXLsTlSyDSdt
++Cc7OdL3FcsctzKemnhaU/cF36QxJdtLg9AvOMXW9dyyVQgjUyBSBm5Q7WQvxvo2
+0xlTqCznX7H3ekeRJFIkGN0ns32e4sPjFdkCAwEAAaAwMBUGCSqGSIb3DQEJBzEI
+EwZhYmMxMjMwFwYJKoZIhvcNAQkCMQoTCEJhcmtiYXVkMA0GCSqGSIb3DQEBBQUA
+A4IBAQAzFIm9X96z7UX4NLHtkvccfBQU7ULm69Q9jiY6eltryrxcBHznBrhLjMBd
+/0BDM+YxuUguVsO9NRSg91dyM4xIdN+TTvCjAeTEvkLx9/q+NKshkfTjkxYwItVr
+9hfwhoau/kwvt8QltFPskq45+A9uxAErXjZnQMk51xmIAdiLffndt7yvuYJwJGfj
+K5AtOKVU/s9S2ujpHoXdTImRuavjU7BxfTkJDs4epUN8/jbBmhU7S5MBTx1nCW6M
+2EN0ZvnAIQXWTpXKiCDP+dmp55ngKwdp+maT9l/6+RMkgaAAfwm5sywFNpeF5vxp
+p9orDJyKZ7NB1M+tNunvjPC457VL
+-----END CERTIFICATE REQUEST-----
