+++ server/sslcerts/server.key
@@ -1,0 +1,27 @@
+-----BEGIN RSA PRIVATE KEY-----
+MIIEpAIBAAKCAQEAqDnZbzkz8m31GdA7/Jhx7STGq0orwq9pMiUP9mVvCNpZViuv
+qZZesCBDP3x7EK0C8HSq03oIK8Uj3WMCCqartbK3OwQdjbT86i6hlNFQXSh1psIW
+e8mncKvvteCzT40vWe+iuUjGyZZ8fo+3FxZb+Gp7tMtP+aufz4MlsOtOXgqa2fTA
+x3KAf6doo7a9TyyBCfymNK83yYvqSKvm+gtzSrv+GLr34F8TtDWyZOQZcuxOVLIN
+J234Jzs50vcVyxy3Mp6aeFpT9wXfpDEl20uD0C84xdb13LJVCCNTIFIGblDtZC/G
++jbTGVOoLOdfsfd6R5EkUiQY3SezfZ7iw+MV2QIDAQABAoIBAQCdLZfj3M2VeHlm
+ifJSW1ku2oqpZl6mw2AFky1r6XAqFGEeKykwyadk5nAOtE8D0LMLWeZ2OvHBe6u7
+4RL6Qqny1a1vfbPjQDSC1djxA/UTZYcvrO528H36vxr0k6LV9DNG06j1d3MMzQc6
+9Sp7bs+1tkApN0FdJKbnq0CfIewR3lCwKa0QT4IOzYxsEQWUBuIWbauD60Xybm5L
+VLn+Qy1S9ikGCtEXOxhkjicyG+NZqbHIzku/oJghkxTYcHNKAQ4Hoc9dDAZiM0Ls
+I/RumFbi0s0zGq6dK+p9+JSy5z0wA0lig3QCuWX5ksAvPcfXpb+0QivLZr7svmjX
+VwIyEmSlAoGBANcO2bBP8LHU2aKrKS0D3i1HAFhnu4MCWbv3uLDN/xcpxuSxXqZ/
+gSv1akwlSxxzH0DpbtkUyhIzf2O+snO8BWf08F70kyUYAn7BtVf78zQrl0BdNBvi
+v+IHrSwo4ZB4FJXdgBjomx7WKOw3g+1BWDX2X403xwz7fCKU+DFoi+oTAoGBAMhA
+kzFvsNse0B4nJ2q0buhF8cf8G4logSxW90/DMs/ePBnbzbhhHv4mwlpx8AzvFyAH
+lJvTrTn9GhbstWW9yEfXoh7wnj5NJfLWVss0/ijXODw2at0wErNxw4fsrrdLHZQI
+9HnpxKOfgzVTaNgSfXKBjqMITwf6lPZjpTPERD3jAoGBALotgQfCLP3G5yKMSvlN
+f+pyrTnbk9m3DzhLHPjzXQMQh780TlyVxZ3ouD1zi7L1dvkbHXuRI97d9krDrHnF
+LGjzA6Uw1rImNwLFb9jbpZBkrGxsQ9BgXP2w55pzSdNoIWocd+kYdjIb9RGKfgvH
+pjYRNGdtAkWa8bx+et9LQtVBAoGAMLx2lluGgW24XP5Oe4u6SlFrjK5wzTpf5bis
+q/7nyOtfWTPDl5IbhuNL/dMpBxzRVfxCUMJtsGVHCjlF1KMhsUprrHgHK0Z/a+B6
+cy5MhQgMKatcysQ+aOnRGq2hv/5KDpM3XMxmDSRvPHU2WQ/28/1QB6GoClsUEfsi
+Y7zqlt0CgYAU163+UF8LC0xnSDOfeKYLaGoqS1U9ER3mrU13yflkxni69uzNJHFb
++Gao6GEMtLP1R/YXG0sECHjdh4AYhKAUQEcTqLhWCYCBd2Wx47oY+zHOs4FGgqMq
+/pAzsE7Khzpu704e1XOJrb953oAtAckzchiBwE6YMHB2FWkfRxiSSg==
+-----END RSA PRIVATE KEY-----
