+++ sky.env-sample
@@ -1,0 +1,5 @@
+AUTH_CLIENT_ID=
+AUTH_CLIENT_SECRET=
+AUTH_REDIRECT_URI=https://localhost:5000/auth/callback
+AUTH_SUBSCRIPTION_KEY=
+PORT=5000
