--- sslcerts/server.crt
@@ -1,19 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDEDCCAfgCCQCOrlP2SsLLnTANBgkqhkiG9w0BAQUFADBKMQswCQYDVQQGEwJV
-UzEXMBUGA1UECBMOU291dGggQ2Fyb2xpbmExEzARBgNVBAcTCkNoYXJsZXN0b24x
-DTALBgNVBAoTBEJFRFMwHhcNMTUxMDE5MjIyNzEyWhcNMTYxMDE4MjIyNzEyWjBK
-MQswCQYDVQQGEwJVUzEXMBUGA1UECBMOU291dGggQ2Fyb2xpbmExEzARBgNVBAcT
-CkNoYXJsZXN0b24xDTALBgNVBAoTBEJFRFMwggEiMA0GCSqGSIb3DQEBAQUAA4IB
-DwAwggEKAoIBAQCxipOI9i/8IYVV3VodJSXWhJHlktM8XTA5s9G5kN0Hrez1RpXQ
-NzlZQrN0iSWymE3k/XkBY2SuEh4GdPFgj3s/QVLe6UUA2gtP83/WKTT9qBzvu1SW
-PMGUkdWn6sEIJkm3hx0Htr52xFm5vzgY6MkUlU+zCgBYPyJCNBwU/pdQU2vJWuwb
-siJ5/kZy3qwNLD2V4gIOhydr64uyaVevafKcElbhsxAjYT2URm3vL0Cvf21WV2jl
-5U0mg/6Zsl2rZCv5NzmvUitgc34AJdmakXd7uSqAyHrKcUz/bSjYnhvryyFylGI+
-3pZHt4je8qpdsKZNhAJiBaToDo3HhG1+ORZ9AgMBAAEwDQYJKoZIhvcNAQEFBQAD
-ggEBAFurmPHrb1KooAehO3BCboEPCg7RdN/O8Mf73lwLVIity5SMXISLiwqq/pZg
-jQe8OfOVmKz4IVu96NxiEqCpyLK8Fm5pWJxMWYi1//ILUUZcTWwq2mTW/++AVBTn
-Pb7NtvvrxCZkuVOzo2iixwM072XRiXYxSgAKdiFZ6LJ+HxP9dsuVH08lx9SiCYXG
-etzTovM2ncxhwVxm2aqDTTUXPjawUwJGXdggfRQ2twoYRTQ/MnaF8dluE5mj4RnX
-LJ8hLKonNGExijz1OJe0LDDv2aShoQYewnd6E/rEHwctB3WzuqrWzu3FsX7MKuOY
-JntzDNdtSk13+xdgJKWe4JmDW9o=
------END CERTIFICATE-----
