--- sslcerts/server.csr
@@ -1,16 +1,0 @@
------BEGIN CERTIFICATE REQUEST-----
-MIICjzCCAXcCAQAwSjELMAkGA1UEBhMCVVMxFzAVBgNVBAgTDlNvdXRoIENhcm9s
-aW5hMRMwEQYDVQQHEwpDaGFybGVzdG9uMQ0wCwYDVQQKEwRCRURTMIIBIjANBgkq
-hkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAsYqTiPYv/CGFVd1aHSUl1oSR5ZLTPF0w
-ObPRuZDdB63s9UaV0Dc5WUKzdIklsphN5P15AWNkrhIeBnTxYI97P0FS3ulFANoL
-T/N/1ik0/agc77tUljzBlJHVp+rBCCZJt4cdB7a+dsRZub84GOjJFJVPswoAWD8i
-QjQcFP6XUFNryVrsG7Iief5Gct6sDSw9leICDocna+uLsmlXr2nynBJW4bMQI2E9
-lEZt7y9Ar39tVldo5eVNJoP+mbJdq2Qr+Tc5r1IrYHN+ACXZmpF3e7kqgMh6ynFM
-/20o2J4b68shcpRiPt6WR7eI3vKqXbCmTYQCYgWk6A6Nx4RtfjkWfQIDAQABoAAw
-DQYJKoZIhvcNAQEFBQADggEBAI+HPnOxdcxRoIffhQZc+eFUk1EykjqlRS79qUYD
-BVcsvUxl4XmPox2eZ2KZ0W93njOIvGpjQYFKwIbznpqFRO1DGUHBR2xP78SspiR0
-yGU647SPWI/UzaWWAaRjQGMnDY3D/z4P/Yqzfthvxr82EJiOXekrX7yHdIxTixc/
-rkNUqMhNqqtdZIeH5+jCK5TXpYd19BZ7cKIu8l9erQtMyiv07iPAWu+v4b3aW8MP
-P/u1W/IBJF87covKbyxLs5eMrczg82ugk0z28OIgOh6JodEW8B7tpuJj6duKuJcB
-6FkttY5HE5R+iJhFUzt1mv0y1jkP+bpxTq6lNtG4gotQ7oM=
------END CERTIFICATE REQUEST-----
