--- sslcerts/server.key
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEowIBAAKCAQEAsYqTiPYv/CGFVd1aHSUl1oSR5ZLTPF0wObPRuZDdB63s9UaV
-0Dc5WUKzdIklsphN5P15AWNkrhIeBnTxYI97P0FS3ulFANoLT/N/1ik0/agc77tU
-ljzBlJHVp+rBCCZJt4cdB7a+dsRZub84GOjJFJVPswoAWD8iQjQcFP6XUFNryVrs
-G7Iief5Gct6sDSw9leICDocna+uLsmlXr2nynBJW4bMQI2E9lEZt7y9Ar39tVldo
-5eVNJoP+mbJdq2Qr+Tc5r1IrYHN+ACXZmpF3e7kqgMh6ynFM/20o2J4b68shcpRi
-Pt6WR7eI3vKqXbCmTYQCYgWk6A6Nx4RtfjkWfQIDAQABAoIBAFYhdmWxmaHl2TvI
-wRj2lp61fdo0GspkIys7M3mk77X0CMsNTUzQHcBR/aGBj4pPHSsEVRn5UDSqkWps
-fVVtBkuvszZ1UIaTQX2uEcyh/RU545/h761bGEatYUuvNapweL7Si6DcZRZVN3pH
-bD2ZYUPdE+c16Ojm6AXaJe2BfYL3xeEj1OELbZQkHd0eFSUltM1+WUoAgT6R1mEU
-qUJfDQ2GZ9R6lYDQkXr1IYlKwTTUSoNhlGqrE+iEKBSWmiFGxMeQOlAcg/VNzjui
-Ajm3OhvhSnFa62qk4M75rOWMQKpPbvJ95DRMq8S1k6MY+HbgBP1rNhNG5qAmLj3o
-Cno3N3kCgYEA6Mo/Sf8TBR2dj7HNmfZkziX8Lb4rB8JZiBSgLwryvMQsbfk708Mn
-EbyyO5xuJoiFuPkYDxYC5ZqNaaCyiJJ0Zff296qrCEUjAR7rUHgrJ7F1ySvOmymV
-bvSi/FqqkfPVnfLhTUplGMxmMfDkKtZZRfLMXzhMRwEnF0DD0uCnQBsCgYEAwz4n
-9+/yjxKDQZcNejiz3qfM6SIUXy7FbO1MHG1z+mZrfrJ/wTxpf2AqT5zH0DlB0nLT
-XHu276BA2ORVNVFbTBTTMFFIl4mkfP8+czQR3Y7QNETH6ewPhCcUeiZvSz2Cjjl2
-9hUsj/lFENdgtf4V+wCdni7accwyobtYO6Zt3UcCgYEAqRvrWBz1uzylbnLTEmq9
-Fkjf5ncEmXFlSwb4ytPgPKMiR43MGWzmn+r9a3KbTmDVC+fgw2zRE+zvl/FGYRJL
-fpvhDFAuSSa4ZCMK/P0fFInFMCwjvfI6fU04bayLxtVktIZxh41o3df20YbWqgvm
-9HqKXw3t05MRy8GfwDy1I0sCgYBbBvClhQaGs+/kNaO/9w3NYliD3TK4Sdsb9OrG
-WjV3aib19AziL+NT0qcHZSY3q2pHzudCOB/LSuDQ/mvHpv201AS+gzcgB7lb2bjp
-Bz7JGFtz8xlts3RsfbfrV14OK0hiPOJ8ytRy75nGx1O8phDjbJ8P7/JCmhyFEQ6L
-CQYx4wKBgB8g13IJVdhAaWoK604GpZOIQOLvA6VFHqUJzJrHwKJGXFpriscNT2ZV
-r5wublKfBBoRlc4g4XDCk3NXOlbLSek/YLC5DY/MMYzoZHW1Q67PGt7unDS4Cz7I
-dfR/g0QDp+mO0sa8K6SotowyIafKWPLnCoCxRha9b3yvgfXBbx5I
------END RSA PRIVATE KEY-----
